Extract save button from brand designer, rename config to brand
