+++ packages/react-brand-designer/src/clipboard-brand-designer.tsx
@@ -1,7 +1,8 @@
 import React from "react";
-import { CourierElement } from "@trycourier/react-elements";
+import { CourierElement, LinkButton } from "@trycourier/react-elements";
 import { BrandConfig } from "./types";
 import { BrandDesigner } from "./brand-designer/brand-designer";
+import styled, { CSSObject } from "styled-components";
 
 /** Copy's Brand Config to Clipboard */
 export const ClipboardBrandDesigner: CourierElement = () => {
@@ -16,18 +17,39 @@ export const ClipboardBrandDesigner: CourierElement = () => {
     setConfig(config);
     setIsCopied(false);
   };
+
   const handleSave = () => {
     navigator.clipboard.writeText(JSON.stringify(config, undefined, 2));
     setIsCopied(true);
   };
 
+  const saveButton = (
+    <CopyButton onClick={handleSave}>
+      {isCopied ? "Copied" : "Copy Snippet"}
+    </CopyButton>
+  );
+
   return (
     <BrandDesigner
-      config={config}
+      brand={config}
       onChange={handleChange}
-      onSave={handleSave}
-      saveButtonText={isCopied ? "Copied" : "Copy Snippet"}
-      disableSaveButton={isCopied}
+      saveButton={saveButton}
     />
   );
 };
+
+const CopyButton = styled(LinkButton)(
+  (): CSSObject => ({
+    fontWeight: 600,
+    fontSize: "12px",
+    lineHeight: "16px",
+    textAlign: "center",
+    textDecorationLine: "underline",
+    color: "#24324B",
+    cursor: "pointer",
+    background: "none",
+    "&:hover": {
+      background: "none",
+    },
+  })
+);
